Move checkbox tree state to special hooks, so as to facilitate reuse
